chore(settings): Move Privacy and Security section to Updates and Help
