+++ lib/l10n/app_fr.arb
@@ -480,8 +480,8 @@
   "applicationSettingsSuggestion": "Apparence, thème, temps et date, etc",
   "privacyAndSecurity": "Sécurité et vie Privée",
   "privacyAndSecuritySettingsSuggestion": "Collecte de données, rapports d'erreur, etc",
-  "updatesAndHelp": "Aide et mise à jour",
-  "updatesAndHelpSettingsSuggestion": "Vérifier les mises à jour et leur historique, etc",
+  "updatesHelpAndPrivacy": "Updates, Help and Privacy",
+  "updatesHelpAndPrivacySettingsSuggestion": "Check for updates, update history, privacy policy, etc",
   "advancedOptions": "Options avancées",
   "advancedOptionsSettingsSuggestion": "Fonctionalitées en Beta, Options de dévelopeur, etc",
   "@@GENERAL": {},
