+++ lib/l10n/app_pl.arb
@@ -504,8 +504,8 @@
   "applicationSettingsSuggestion": "Appearance, theme, date and time, etc",
   "privacyAndSecurity": "Privacy and Security",
   "privacyAndSecuritySettingsSuggestion": "Data collection, error reporting, etc",
-  "updatesAndHelp": "Updates and Help",
-  "updatesAndHelpSettingsSuggestion": "Check for updates, update history, etc",
+  "updatesHelpAndPrivacy": "Updates, Help and Privacy",
+  "updatesHelpAndPrivacySettingsSuggestion": "Check for updates, update history, privacy policy, etc",
   "advancedOptions": "Advanced Options",
   "advancedOptionsSettingsSuggestion": "Funcionalidades em Beta, Opções de Desenvolvedor, etc",
   "@@GENERAL": {},
