+++ lib/l10n/app_pt.arb
@@ -504,8 +504,8 @@
   "applicationSettingsSuggestion": "Aparência, idioma, etc",
   "privacyAndSecurity": "Privacidade e Segurança",
   "privacyAndSecuritySettingsSuggestion": "Coleta de dados, relatórios de erros, etc",
-  "updatesAndHelp": "Atualizações e Ajuda",
-  "updatesAndHelpSettingsSuggestion": "Verificar atualizações, histórico de atualizações, etc",
+  "updatesHelpAndPrivacy": "Atualizações, Ajuda e Privacidade",
+  "updatesHelpAndPrivacySettingsSuggestion": "Procure por atualizações, histórico de mudanças, política de privacidade, etc",
   "advancedOptions": "Opções Avançadas",
   "advancedOptionsSettingsSuggestion": "Funcionalidades em Beta, Opções de Desenvolvedor, etc",
   "@@GENERAL": {},
